requirements.txt
flask==2.2.2
flask-cors==4.0.0
google-generativeai==0.3.1
python-dotenv==1.0.0
beautifulsoup4==4.12.2
requests==2.31.0
PyPDF2==3.0.1
python-docx==1.0.1
selenium==4.15.2
webdriver-manager==4.0.1
gunicorn==21.2.0
Werkzeug==2.3.7
